Thanks to Przemek Wojnow.
Current version of this function has wrong regexp,
which returns true when last subpackage is the same
as current package. For example:
--- Plik.java ---
package tree;
public class Plik {
/* Importing this class will fail, because it's
in package javax.swing.tree */
DefaultMutableTreeNode node;
}

The class passed to this function is always fully
qualified, so its full package name should be
compared to the current.


git-svn-id: https://jdee.svn.sourceforge.net/svnroot/jdee/branches/2.4.0@132 381a1f10-3442-0410-997e-ef331cd04102