I found two bugs when using `jde-import-all':
1. Classes not being imported.
2. Additional lines inserted when importing classes.

The first bug introduced in 181 when adding this to jde.el

  (autoload 'wisent-java-default-setup "wisent-java" "Hook run to setup Semantic in `java-mode'." nil nil)

which comes from:

  http://www.mail-archive.com/jde@sunsite.dk/msg07334.html

Looks like the 1.5 implementation is buggy as the URL states it might be.  Perhaps it will work better in the newer cedet, but for now it doesn't make sense to break a major feature like `jde-import-all'.

The second bug was introduced in 183 and inserts lines around inserted imports.


git-svn-id: https://jdee.svn.sourceforge.net/svnroot/jdee/branches/2.4.0@191 381a1f10-3442-0410-997e-ef331cd04102